refactor

+++ Mai/VideoManager.swift
@@ -19,11 +19,13 @@ import RxReachability
 extension VideoManager {
     
     private enum K {
-        static let cacheSizeLimit = 100 * (1 << 20)  // 100 MB
+        static let cacheSizeLimit: UInt64 = 100 * (1 << 20)  // 100 MB
 
         static let cacheDir = Path.userMovies + "Mai" + ".cache"
         static let likeDir = Path.userMovies + "Mai" + "like"
         static let dislikeDir = Path.userMovies + "Mai" + ".dislike"
+        
+        static let downloadInterval: TimeInterval = 5
         
         static let apiHost = "animeloop.org"
         static let baseURL = "https://animeloop.org/api/v2"
@@ -35,46 +37,42 @@ extension VideoManager {
     }
 }
 
+extension VideoManager {
+ 
+    var async: Observable<VideoManager> {
+        return Observable.just(self).observeOn(ConcurrentDispatchQueueScheduler(qos: .userInitiated))
+    }
+}
+
 final class VideoManager {
 
     private let disposeBag = DisposeBag()
     private let ioQueue = DispatchQueue(label: UUID().uuidString)
-    private let reachability = Reachability(hostname: K.apiHost)
-    private var isFetchingEnabled = true
+    private let reachability = Reachability(hostname: K.apiHost)!
+    
+    private var cacheTotalSize: UInt64 = 0
+    
     private var fetchDisposable: Disposable?
-    private weak var retryTimer: Timer?
 
     // MARK: - Init
     private init() {
         createDirIfNeeded()
         copyDefaultVideoIfNeeded()
         cleanCacheDirIfNeeded()
-
-        reachability?.rx
-            .status
-            .distinctUntilChanged()
-            .bind { [weak self] c in
-                guard let self = self else { return }
-                Logger.debug("API reachability changed", "connection: \(c)")
-                if c == .none {
-                    Logger.warn("API is unreachable, stop fetching new videos")
-                    self.fetchDisposable = nil
-                    self.retryTimer?.invalidate()
-                } else {
-                    Logger.info("API is reachable, start to fetch a new video")
-                    if self.fetchDisposable == nil && self.retryTimer == nil {
-                        self.fetch()
-                    }
-                }
-            }
-            .disposed(by: disposeBag)
+        
+        do {
+            try reachability.startNotifier()
+        } catch let e {
+            Logger.error("Failed to set up reachability notifier", e)
+        }
 
         EventBus.onlyLiked
             .bind { [weak self] flag in
+                guard let self = self else { return }
                 if flag {
-                    self?.stopTryingFetching()
+                    self.stopFetching()
                 } else {
-                    self?.fetchIfPossible()
+                    self.startFetching()
                 }
             }
             .disposed(by: disposeBag)
@@ -84,12 +82,12 @@ final class VideoManager {
 
     private func createDirIfNeeded() {
         for p in [K.cacheDir, K.likeDir, K.dislikeDir] {
-            if !p.exists {
-                do {
-                    try p.createDirectory()
-                } catch let err {
-                    Logger.error("Failed to create directory", p, err)
-                }
+            if p.exists { continue }
+            
+            do {
+                try p.createDirectory()
+            } catch let err {
+                Logger.error("Failed to create directory", p, err)
             }
         }
     }
@@ -98,53 +96,45 @@ final class VideoManager {
         if let path = Bundle.main.path(forResource: "5bbadd3466e1f3205b7e4e98", ofType: "mp4") {
             let from = Path(path)
             let to = K.cacheDir + from.fileName
-            if !to.exists {
-                do {
-                    try from.moveFile(to: to)
-                } catch let err {
-                    Logger.error("Failed to copy default video", err)
-                }
+            if to.exists { return }
+            
+            do {
+                try from.moveFile(to: to)
+            } catch let err {
+                Logger.error("Failed to copy default video", err)
             }
         }
     }
 
     private func cleanCacheDirIfNeeded() {
-        ioQueue.async {
-            var totalSize = K.cacheDir.children().reduce(into: 0, { $0 += ($1.fileSize ?? 0) })
+        ioQueue.async { [weak self] in
+            guard let self = self else { return }
+            self.cacheTotalSize = K.cacheDir.children().reduce(into: 0, { $0 += ($1.fileSize ?? 0) })
             var files = K.cacheDir.children()
                 .filter {
                     $0.pathExtension == "mp4"
                 }
                 .sorted(by: K.descendingByCreationDate)
-                .map({ $0 })
 
-            var deleted: [Path] = []
-            while totalSize > K.cacheSizeLimit {
+            while self.cacheTotalSize > K.cacheSizeLimit {
                 guard let path = files.popLast() else {
-                    Logger.error("No file???")
-                    preconditionFailure("No file???")
+                    Logger.error("Cache folder size is \(self.cacheTotalSize) but no mp4 file???")
+                    return
                 }
+                
                 do {
                     guard let fileSize = path.fileSize else { continue }
-                    totalSize -= fileSize
+                    self.cacheTotalSize -= fileSize
                     try path.deleteFile()
-                    deleted.append(path)
                 } catch let err {
                     Logger.error("Failed to delete file", path, err)
                 }
             }
-
-            if !deleted.isEmpty {
-                Logger.cheer("Cache dir has been cleaned", deleted)
-            }
-            DispatchQueue.global().asyncAfter(deadline: .now() + 15) { [weak self] in
-                self?.cleanCacheDirIfNeeded()
-            }
         }
     }
 
-    // MARK: - Properties
-    var allCachedVideo: [URL] {
+    // MARK: - All videos
+    var allCachedVideos: [URL] {
         return ioQueue.sync {
             return K.cacheDir
                 .children()
@@ -163,102 +153,126 @@ final class VideoManager {
                 .map({ $0.url })
         }
     }
+    
+    var allDislikedVideos: [URL] {
+        return ioQueue.sync {
+            return K.dislikeDir
+                .children()
+                .filter({ $0.pathExtension == "mp4" })
+                .sorted(by: K.descendingByCreationDate)
+                .map({ $0.url })
+        }
+    }
 
-    // MARK: Like
+    // MARK: Like & Dislike
     func like(_ url: URL) {
-        ioQueue.sync {
-            if let path = Path(url: url) {
-                let dest = K.likeDir + path.fileName
-                guard !dest.exists else { return }
-                do {
-                    Logger.info("Copy the video to like dir", dest)
-                    try path.copyFile(to: dest)
-                } catch let err {
-                    Logger.error("Failed to move file to like dir", dest, err)
-                }
+        ioQueue.async {
+            guard let path = Path(url: url) else { return }
+            
+            let dest = K.likeDir + path.fileName
+            if dest.exists { return }
+            do {
+                try path.copyFile(to: dest)
+            } catch let err {
+                Logger.error("Failed to copy file to like dir", err)
             }
         }
     }
 
     func dislike(_ url: URL) {
-        ioQueue.sync {
-            if let path = Path(url: url), path.exists {
-                let dest = K.dislikeDir + path.fileName
-                if dest.exists {
-                    do {
-                        try path.deleteFile()
-                    } catch let err {
-                        Logger.error("Failed to delete file", path, err)
-                    }
-                    return
-                }
-                do {
-                    Logger.info("Move the video to dislike dir", dest)
-                    try path.moveFile(to: dest)
-                } catch let err {
-                    Logger.error("Failed to move file to dislike dir", dest, err)
-                }
+        ioQueue.async {
+            guard let path = Path(url: url), path.exists else {
+                return
+            }
+            let dest = K.dislikeDir + path.fileName
+            if dest.exists { return }
+            
+            do {
+                try path.moveFile(to: dest)
+                self.cacheTotalSize -= dest.fileSize ?? 0
+            } catch let err {
+                Logger.error("Failed to move file to dislike dir", dest, err)
             }
         }
     }
 
-    // MARK: Fetch
+    // MARK: Download
     private func fetch() {
-        guard isFetchingEnabled else { return }
-        Logger.info("Fetching...")
-
-        fetchDisposable = RxAlamofire
-            .json(.get,
-                  K.baseURL + "/rand/loop",
-                  parameters: ["full": true, "limit": 1]
-            )
-            .flatMap { (obj) -> Observable<(String, Data)> in
-                if let url = JSON(obj)["data"][0]["files"]["mp4_1080p"].string {
-                    return RxAlamofire.data(.get, url).map { (url, $0) }
+        fetchDisposable?.dispose()
+        fetchDisposable = Observable<Int>.timer(0, period: K.downloadInterval, scheduler: MainScheduler.asyncInstance)
+            .flatMapLatest({ (_) -> Observable<Void> in
+                
+                guard self.reachability.isReachable else {
+                    return .empty()
                 }
-                return Observable.empty()
-            }
-            .subscribe(onNext: { [weak self] (url, data) in
-                let fileName = Path(url).fileName
-                let dest = K.cacheDir + fileName
-                do {
-                    try self?.ioQueue.sync {
-                        if dest.exists {
-                            Logger.info("A new video has been downloaded but we already have it", fileName)
-                        } else {
-                            Logger.cheer("A new video has been downloaded and written to disk", fileName)
-                            try data.write(to: dest)
+                
+                return RxAlamofire
+                    .json(.get,
+                          K.baseURL + "/rand/loop",
+                          parameters: ["full": true, "limit": 1]
+                    )
+                    .flatMap { (obj) -> Observable<Void> in
+                        guard
+                            let path = JSON(obj)["data"][0]["files"]["mp4_1080p"].string,
+                            let url = URL(string: path)
+                            else {
+                                return Observable.empty()
                         }
+                        
+                        if (self.allCachedVideos + self.allDislikedVideos).contains(where: { $0.lastPathComponent == url.lastPathComponent }) {
+                            return Observable.empty()
+                        }
+                        
+                        return Observable<Void>.create({ (o) -> Disposable in
+                            let request = Alamofire.download(url)
+                            request.response(completionHandler: { (res) in
+                                if let err = res.error {
+                                    o.onError(err)
+                                    return
+                                }
+                                guard let url = res.temporaryURL else {
+                                    return
+                                }
+                                
+                                self.ioQueue.async {
+                                    let tmpPath = Path(url: url)!
+                                    let fileName = tmpPath.fileName
+                                    let dest = K.cacheDir + fileName
+                                    
+                                    if dest.exists { return }
+                                    
+                                    do {
+                                        try tmpPath.moveFile(to: dest)
+                                        self.cacheTotalSize += dest.fileSize ?? 0
+                                    } catch let e {
+                                        Logger.error("Failed to download video", e)
+                                    }
+                                    DispatchQueue.main.async {
+                                        EventBus.newVideo.accept(url)
+                                    }
+                                }
+                            })
+                            
+                            return Disposables.create {
+                                request.cancel()
+                            }
+                        })
                     }
-                    EventBus.newVideo.accept(dest.url)
-                    self?.retryTimer = Timer.scheduledTimer(withTimeInterval: 5, repeats: false, block: { (_) in
-                        self?.fetch()
-                    })
-                } catch let err {
-                    Logger.error("Failed to write video to disk", dest, err)
-                }
-            }, onError: { [weak self] (err) in
+            })
+            .subscribe(onNext: { _ in
+            }, onError: { (err) in
                 Logger.error("Failed to download video", err)
-                self?.retryTimer = Timer.scheduledTimer(withTimeInterval: 5, repeats: false, block: { (_) in
-                    self?.fetch()
-                })
             })
     }
 
-    func fetchIfPossible() {
-        Logger.info("Fetch If Possible")
-        try? reachability?.startNotifier()
-        isFetchingEnabled = true
-        if fetchDisposable == nil && retryTimer == nil {
+    func startFetching() {
+        if fetchDisposable == nil {
             fetch()
         }
     }
 
-    func stopTryingFetching() {
-        Logger.info("Stop Trying Fetching")
-        reachability?.stopNotifier()
-        isFetchingEnabled = false
+    func stopFetching() {
+        fetchDisposable?.dispose()
         fetchDisposable = nil
-        retryTimer?.invalidate()
     }
 }
